+++ Jenkinsfile
@@ -182,9 +182,7 @@ def build_casync_release(String channel_name, def is_release) {
   def extra_env = is_release ? "RELEASE=1" : ""
 
   return deviceStage("build casync", "tici-needs-can", [], [
-    ["build", "${extra_env} BUILD_DIR=/data/openpilot CASYNC_DIR=/data/casync/openpilot $SOURCE_DIR/release/create_casync_build.sh"],
-    ["create manifest", "$SOURCE_DIR/release/create_release_manifest.py /data/openpilot /data/manifest.json && cat /data/manifest.json"],
-    ["upload and cleanup", "PYTHONWARNINGS=ignore $SOURCE_DIR/release/upload_casync_release.py /data/casync && rm -rf /data/casync"],
+    ["build", "${extra_env} BUILD_DIR=/data/openpilot CASYNC_DIR=/data/casync$SOURCE_DIR/release/build_casync_release.sh"],
   ])
 }
